move and adjust original API
