+++ setup.py
@@ -1,18 +1,19 @@
 from setuptools import setup
+from setuptools import find_packages
 
-setup(name='mpl_finance',
+setup(name='mplfinance',
       version='0.10.0',
       author='MPL Developers',
       author_email='matplotlib-users@python.org',
-      py_modules=['mpl_finance'],
+      py_modules=['mplfinance'],
       description='Finance plots using matplotlib',
-      url='http://github.com/matplotlib/mpl_finance',
+      url='http://github.com/matplotlib/mplfinance',
       platforms='Cross platform (Linux, Mac OSX, Windows)',
       install_requires=['matplotlib'],
       license="BSD",
+      package_dir={'': 'src'},
+      packages=find_packages(where='src'),
       classifiers=['Development Status :: 4 - Beta',
-                   'Programming Language :: Python :: 2',
-                   'Programming Language :: Python :: 2.7',
                    'Programming Language :: Python :: 3',
                    'Programming Language :: Python :: 3.3',
                    'Programming Language :: Python :: 3.4',
